+++ allennlp/data/fields/field.py
@@ -4,6 +4,7 @@ from typing import Dict, Generic, List, TypeVar
 import torch
 
 from allennlp.data.vocabulary import Vocabulary
+from allennlp.common.util import JsonDict
 
 DataArray = TypeVar(
     "DataArray", torch.Tensor, Dict[str, torch.Tensor], Dict[str, Dict[str, torch.Tensor]]
@@ -53,25 +54,20 @@ class Field(Generic[DataArray]):
         """
         pass
 
-    def human_readable_dict(self):
+    def human_readable_dict(self) -> JsonDict:
         """
         This function facilitate saving formated instances to json files for human readability,
         use case includes example-based explanation, where it's better to have a output file
-        rather than printing or logging.
+        rather than printing or logging. The output will be a dictionary, which contains a
+        default key/name of the field, then a structured representation of the field.
 
-        For example, - if the field is LabelField, then we just output, field.label
-                     - if the field is TextField, then we just output, field.tokens
-                       (preferrably un-numericalized tokens)
+        For example, - if the field is LabelField, then we just output,
+                        {"label": field.label (string)}
+                     - if the field is TextField, then we just output,
+                        {"tokens": field.tokens (preferrably un-numericalized tokens)}
+
         Since this is hard to deal with in higher level usage -- e.g. judging instance contains which fields
         and how to convert -- it's better to do it in the lower level.
-
-        # Parameters
-        human_readable : `bool`
-                        a flag to control the json output of the field.
-                        For example, If the field is TensorField, then it makes less sense to
-                        output it for human readabilty. But still, if the user
-                        really want to output it, we still allows it. In contrast, if it's label field,
-                        then, simply outputing it seems fine.
         """
         raise NotImplementedError